:recycle: [refactor] : resolved conflicts
