+++ src/app/_types/response.ts
@@ -7,12 +7,26 @@ interface ResponsePortfolioDetails {
 
 interface ResponseRecruitingMain {
   user: UserProfile;
-  pagable: Pagable;
+  pageable: Pageable;
   recrutingList: RecruitListItem[];
+}
+
+interface ResponsePortfolioMain {
+  user: UserProfile;
+  pageable: Pageable;
+  hotArticleList: HotArticle[];
+  articleList: PortfolioArticle[];
+  recommendArticleList: RecommendArticle[];
+}
+
+interface ResponsePortfolioMainDetail {
+  user: UserProfile;
+  recruit: RecruitDetailItem;
+  comments: CommentItem[];
 }
 
 interface ResponseRecruitingDetail {
   user: UserProfile;
   recruit: RecruitDetailItem;
-  comments: Comment[];
+  comments: CommentItem[];
 }
