+++ src/app/components/main/common/DetailCategories.tsx
@@ -4,7 +4,7 @@ interface DetailCategoriesProps {
 
 const DetailCategories = ({ category }: DetailCategoriesProps) => {
   return (
-    <div className="flex mr-[20px] w-[105px] max-w-[120px] px-2 h-9 pt-1  bg-gradient-to-r from-[#4EF4FF] to-[#608CFF] rounded-3xl justify-center items-center gap-5">
+    <div className="flex mr-[20px] w-[105px] max-w-[120px] px-2 h-9 pt-1 bg-gradient-to-r from-[#00E0EE] to-[#517EF3] rounded-3xl justify-center items-center gap-5">
       <div># {category}</div>
     </div>
   );
