+++ src/app/components/main/containers/HotPortfolioContainer.tsx
@@ -1,12 +1,12 @@
 'use client';
 
-import HotPortfolioItem from '../portfolio/HotPortfolioItem';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation } from 'swiper/modules';
 import 'swiper/css';
 import 'swiper/css/navigation';
 import Icons from 'components/common/Icons';
 import { bigLeftAngle, bigRightAngle } from 'ui/IconsPath';
+import HotPortfolioItem from '../HotPortfolioItem';
 
 interface HotPortfolioContainerProps {
   data: HotArticle[];
