+++ src/app/components/main/containers/RecommendContainer.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import { LOGGEDIN_SECTION_HEADINGS } from 'constants/main';
-import PortfolioItem from '../portfolio/PortfolioItem';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Autoplay } from 'swiper/modules';
 import 'swiper/css';
+import PortfolioItem from '../PortfolioItem';
 
 interface RecommendContainerProps {
   data: RecommendArticle[];
